Change closed_loop param format for automatic evaluation in CARLA

+++ global_planning/include/global_planning_alg_node.h
@@ -73,7 +73,7 @@ private:
   tf::TransformBroadcaster broadcaster_;
   tf::TransformListener listener_;
   Graph *graph_;
-  std::vector<double> closed_loop_;
+  std::vector<int> closed_loop_;
 
   // [publisher attributes]
   ros::Publisher marker_pub_;
